fix: Assign IPV4 address automatically & fix IAM
The issue was that the newly launched instances didn't have public IPV4
addresses thus they weren't accessible on the network. The solution was
to associate IPV4 addresses automatically on the subnet level when a new
instance launch initiated.

Another problem was a missing IAM instance profile, so the launch
template didn't recognize the IAM role as a valid setup. The solution
was to create an IAM instance profile and associate it with the
previously created IAM role.